pyproject.toml
[project]
name = "playwright_automation_python"
version = "0.1.0"
description = "Playwright automation practice with python."
authors = [
    {name = "Sabbir", email = 'sabbir.ul.alam.sabbir@gmail.com'}
]
package-mode = false
#readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-xdist (>=3.6.1,<4.0.0)",
    "pytest-rerunfailures (>=15.1,<16.0)",
    "pytest-playwright (>=0.7.0)"
    ]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
tox = "^4.26.0"
pytest-playwright = "^0.7.0"
flake8 = "^7.2.0"

